move infrastructure into separate project
