+++ src/Infrastruture/IBus.cs
@@ -1,4 +1,4 @@
-﻿namespace Registration.infrastructure
+﻿namespace Infrastruture
 {
     public interface IBus : IPublish, ISubscribe
     {
