+++ src/Infrastruture/IEvent.cs
@@ -1,3 +1,3 @@
-﻿namespace Registration.infrastructure{
+﻿namespace Infrastruture{
     public interface IEvent : IMessage { }
 }
