+++ src/Infrastruture/IHandle.cs
@@ -1,4 +1,4 @@
-﻿namespace Registration.infrastructure{
+﻿namespace Infrastruture{
     public interface IHandle<in TEvent> where TEvent : IEvent
     {
         void Handle(TEvent @event);
