+++ src/Infrastruture/IHandleCommand.cs
@@ -1,4 +1,4 @@
-﻿namespace Registration.infrastructure{
+﻿namespace Infrastruture{
     public interface IHandleCommand<in TCommand> where TCommand : ICommand
     {
         bool Handle(TCommand cmd);
