+++ src/Infrastruture/IMessage.cs
@@ -1,0 +1,3 @@
+﻿namespace Infrastruture{
+    public interface IMessage { }
+}
