+++ src/Infrastruture/IRepository.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Registration.infrastructure {
+namespace Infrastruture {
     public interface IRepository {
         void Save(IEventSource source);
         T Load<T>(Guid id) where T : class, IEventSource;
