+++ src/Registration/Application/Bootstrap.cs
@@ -3,13 +3,13 @@ using System.Net;
 using System.Text;
 using EventStore.ClientAPI;
 using EventStore.ClientAPI.SystemData;
+using Infrastruture;
 using Newtonsoft.Json;
 using Registration.Blueprint.Commands;
 using Registration.Blueprint.Events;
 using Registration.Blueprint.ReadModels;
-using Registration.CommandHandlers;
-using Registration.EventReaders;
-using Registration.infrastructure;
+using Registration.Components.CommandHandlers;
+using Registration.Components.EventReaders;
 
 namespace Registration.Application
 {
@@ -28,25 +28,30 @@ namespace Registration.Application
                 .Build();
             var conn = EventStoreConnection.Create(settings, IPEndPoint.Parse("127.0.0.1:1113"));
             conn.ConnectAsync().Wait();
-
-            var mainBus = new SimpleBus();
+            
+            //configure read side
             var eventBus = new SimpleBus();
-
             var repo = new SimpleRepo(conn);
-            var pump = new EventPump(conn, eventBus, repo.Deserialize);
-
+            
             var userRm = new RegisteredUsers();
             eventBus.Subscribe<UserRegistered>(userRm);
             eventBus.Subscribe<NameChanged>(userRm);
+
+            var pump = new EventPump(conn, eventBus, repo.Deserialize);
+
+            //configure write side
+            var mainBus = new SimpleBus();
 
             var userSvc = new UserSvc(repo);
             mainBus.Subscribe<RegisterUser>(userSvc);
             mainBus.Subscribe<ChangeName>(userSvc);
 
-            pump.Start();
-
+            //application wire up
             app.CommandPublisher = mainBus;
             app.GetUsers = () => (IRegisteredUsers)userRm;
+            
+            //start 
+            pump.Start();
         }
     }
 }
