+++ src/Registration/Registration.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29020.237
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Registration", "Registration.csproj", "{E8286303-C581-4080-A8B7-F54657265794}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Registration", "Registration.csproj", "{E8286303-C581-4080-A8B7-F54657265794}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastruture", "..\Infrastruture\Infrastruture.csproj", "{AF5DE214-12A1-4379-A58B-EDCD4A9F99B6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{E8286303-C581-4080-A8B7-F54657265794}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E8286303-C581-4080-A8B7-F54657265794}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E8286303-C581-4080-A8B7-F54657265794}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AF5DE214-12A1-4379-A58B-EDCD4A9F99B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AF5DE214-12A1-4379-A58B-EDCD4A9F99B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AF5DE214-12A1-4379-A58B-EDCD4A9F99B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AF5DE214-12A1-4379-A58B-EDCD4A9F99B6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
