--- src/Registration/infrastructure/IMessage.cs
@@ -1,3 +1,0 @@
-﻿namespace Registration.infrastructure{
-    public interface IMessage { }
-}
